feature(frontend): remove old implementation
