--- frontend/src/lib/components/flows/CopyFirstStepSchema.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-	import { Button } from 'flowbite-svelte'
-	import { isCopyFirstStepSchemaDisabled } from './flowState'
-	import { copyFirstStepSchema } from './flowStore'
-</script>
-
-<Button
-	color="light"
-	size="xs"
-	disabled={$isCopyFirstStepSchemaDisabled}
-	on:click={copyFirstStepSchema}
->
-	Copy from first step schema
-</Button>
