+++ frontend/src/lib/components/flows/content/FlowEditorPanel.svelte
