+++ frontend/src/lib/components/flows/content/FlowModuleHeader.svelte
@@ -1,4 +1,7 @@
 <script lang="ts">
+	import { isEmptyFlowModule } from '$lib/components/flows/flowStateUtils'
+	import Modal from '$lib/components/Modal.svelte'
+
 	import type { FlowModule } from '$lib/gen'
 	import { getScriptByPath } from '$lib/utils'
 	import {
@@ -15,23 +18,18 @@
 	import { Highlight } from 'svelte-highlight'
 	import python from 'svelte-highlight/languages/python'
 	import typescript from 'svelte-highlight/languages/typescript'
-	import IconedPath from '../IconedPath.svelte'
-	import Modal from '../Modal.svelte'
-	import { isEmptyFlowModule } from './flowStateUtils'
-	import { stepOpened } from './stepOpenedStore'
 
-	export let indexes: number[]
-	export let mod: FlowModule
+	export let module: FlowModule
 
-	$: shouldPick = isEmptyFlowModule(mod)
+	$: shouldPick = isEmptyFlowModule(module)
 
 	let modalViewer: Modal
 	let modalViewerContent = ''
 	let modalViewerLanguage: 'deno' | 'python3' = 'deno'
 
 	async function viewCode() {
-		if (mod.value.type == 'script') {
-			const { content, language } = await getScriptByPath(mod.value.path!)
+		if (module.value.type == 'script') {
+			const { content, language } = await getScriptByPath(module.value.path!)
 			modalViewerContent = content
 			modalViewerLanguage = language
 			modalViewer.openModal()
@@ -41,34 +39,10 @@
 	}
 
 	const dispatch = createEventDispatcher()
-	$: opened = $stepOpened === String(indexes.join('-'))
 </script>
 
-<div>
-	<h3 class="text-sm font-bold text-gray-900 flex flex-col">
-		<span class="text-xs shrink">
-			{#if 'path' in mod.value && mod.value.path}
-				<IconedPath path={mod.value.path} />
-			{:else if 'language' in mod.value && mod.value.language}
-				Inline {mod.value.language}
-			{:else}
-				Select a script
-			{/if}
-		</span>
-		{#if ('path' in mod.value && mod.value.path) || ('language' in mod.value && mod.value.language)}
-			<input
-				on:click|stopPropagation={() => undefined}
-				class="overflow-x-auto"
-				type="text"
-				bind:value={mod.summary}
-				placeholder="Summary"
-			/>
-		{/if}
-	</h3>
-</div>
-
 <div class="flex flex-row space-x-2" on:click|stopPropagation={() => undefined}>
-	{#if mod.value.type === 'script' && !shouldPick}
+	{#if module.value.type === 'script' && !shouldPick}
 		<Button size="xs" color="alternative" on:click={() => dispatch('fork')}>
 			<Icon data={faCodeBranch} class="mr-2" />
 			Fork
@@ -79,7 +53,7 @@
 		</Button>
 	{/if}
 
-	{#if mod.value.type === 'rawscript' && !shouldPick}
+	{#if module.value.type === 'rawscript' && !shouldPick}
 		<Button size="xs" color="alternative" on:click={() => dispatch('createScriptFromInlineScript')}>
 			<Icon data={faSave} class="mr-2" />
 			Save to workspace
@@ -95,23 +69,10 @@
 		<Icon data={faTrashAlt} class="mr-2" />
 		Remove step
 	</Button>
-	{#if opened}
-		<Button size="xs" color="dark" on:click={() => stepOpened.update(() => undefined)}>
-			<Icon data={faClose} />
-		</Button>
-	{:else}
-		<Button
-			size="xs"
-			color="light"
-			on:click={() => stepOpened.update(() => String(indexes.join('-')))}
-		>
-			<Icon data={faArrowDown} />
-		</Button>
-	{/if}
 </div>
 
 <Modal bind:this={modalViewer}>
-	<div slot="title">Script {'path' in mod?.value ? mod?.value.path : ''}</div>
+	<div slot="title">Script {'path' in module?.value ? module?.value.path : ''}</div>
 	<div slot="content">
 		{#if modalViewerLanguage === 'python3'}
 			<Highlight language={python} code={modalViewerContent} />
