+++ frontend/src/lib/components/flows/utils.ts
@@ -122,8 +122,9 @@ export function getDefaultExpr(
 	previousExpr?: string
 ) {
 	const expr = previousExpr ?? `previous_result.${key}`
-	return `import { previous_result, flow_input, step, variable, resource, params } from 'windmill${importPath ? `@${importPath}` : ''
-		}'
+	return `import { previous_result, flow_input, step, variable, resource, params } from 'windmill${
+		importPath ? `@${importPath}` : ''
+	}'
 
 ${expr}`
 }
@@ -165,4 +166,15 @@ export function codeToStaticTemplate(code?: string): string | undefined {
 		}
 	}
 	return undefined
+}
+
+export function getIndexes(parentIndex: number | undefined, childIndex: number): number[] {
+	const indexes: number[] = []
+
+	if (parentIndex !== undefined) {
+		indexes.push(parentIndex)
+	}
+	indexes.push(childIndex)
+
+	return indexes
 }
