--- frontend/src/lib/components/flows_neo/utils.ts
@@ -1,10 +1,0 @@
-export function getIndexes(parentIndex: number | undefined, childIndex: number): number[] {
-	const indexes: number[] = []
-
-	if (parentIndex !== undefined) {
-		indexes.push(parentIndex)
-	}
-	indexes.push(childIndex)
-
-	return indexes
-}
